+++ feature/wiki/src/main/kotlin/com/github/felipehjcosta/marvelapp/wiki/presentation/CharacterItemViewModel.kt
@@ -1,3 +1,3 @@
 package com.github.felipehjcosta.marvelapp.wiki.presentation
 
-class CharacterItemViewModel(val id: Int, val name: String, val image: String)
+class CharacterItemViewModel(val id: Long, val name: String, val image: String)
